[Refactor] 위치 설정 모달 검색 로직 및 기본 위치값 분리 (#128)
